Improved SelectMany() against value collections. Closes GH-2704. Closes GH-2706
