+++ src/Marten/Linq/Members/ValueCollections/IValueCollectionMember.cs
@@ -1,0 +1,6 @@
+namespace Marten.Linq.Members.ValueCollections;
+
+internal interface IValueCollectionMember : IQueryableMemberCollection
+{
+    IQueryableMember Element { get; }
+}
