+++ src/Marten/Linq/SqlGeneration/ScalarSelectClause.cs
@@ -61,7 +61,7 @@ internal class ScalarSelectClause<T>: ISelectClause, ISelector<T>, IScalarSelect
     {
         sql.Append("select ");
         sql.Append(MemberName);
-        sql.Append(" from ");
+        sql.Append(" as data from ");
         sql.Append(FromObject);
         sql.Append(" as d");
     }
