+++ src/Marten/Marten.csproj
@@ -23,31 +23,31 @@
     </ItemGroup>
 
     <ItemGroup Condition="'$(TargetFramework)' == 'net6.0'">
-        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="6.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="6.0.0" />
     </ItemGroup>
 
     <ItemGroup Condition="'$(TargetFramework)' == 'net7.0'">
-        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="7.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="7.0.0" />
     </ItemGroup>
 
     <ItemGroup Condition="'$(TargetFramework)' == 'net8.0'">
-        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="8.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="8.0.0" />
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="JasperFx.Core" Version="1.5.1"/>
-        <PackageReference Include="JasperFx.CodeGeneration" Version="3.4.0"/>
-        <PackageReference Include="JasperFx.RuntimeCompiler" Version="3.4.0"/>
-        <PackageReference Include="Newtonsoft.Json" Version="13.0.3"/>
+        <PackageReference Include="JasperFx.Core" Version="1.5.1" />
+        <PackageReference Include="JasperFx.CodeGeneration" Version="3.4.0" />
+        <PackageReference Include="JasperFx.RuntimeCompiler" Version="3.4.0" />
+        <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
         <!-- This is forced by Npgsql peer dependency -->
-        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.0"/>
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.0"/>
+        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.0" />
+        <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.0" />
         <!-- This is forced by Npgsql peer dependency -->
-        <PackageReference Include="Npgsql.Json.NET" Version="8.0.0"/>
-        <PackageReference Include="Remotion.Linq" Version="2.2.0"/>
-        <PackageReference Include="System.Threading.Tasks.Dataflow" Version="8.0.0"/>
-        <PackageReference Include="Weasel.Postgresql" Version="7.0.0-alpha.2"/>
-        <PackageReference Include="System.Text.Json" Version="8.0.0"/>
+        <PackageReference Include="Npgsql.Json.NET" Version="8.0.0" />
+        <PackageReference Include="Remotion.Linq" Version="2.2.0" />
+        <PackageReference Include="System.Threading.Tasks.Dataflow" Version="8.0.0" />
+        <PackageReference Include="Weasel.Postgresql" Version="7.0.0-alpha.2" />
+        <PackageReference Include="System.Text.Json" Version="8.0.0" />
     </ItemGroup>
 
     <!--SourceLink specific settings-->
@@ -63,7 +63,7 @@
         <SymbolPackageFormat>snupkg</SymbolPackageFormat>
     </PropertyGroup>
     <ItemGroup>
-        <PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All"/>
+        <PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All" />
     </ItemGroup>
     <Import Project="../../Analysis.Build.props" />
 </Project>
